Move fmt::format to fmt/format.h
