+++ include/fmt/core.h
@@ -1,1 +1,5 @@
+// This file is only provided for compatibility an may be removed in future
+// versions. Use fmt/base.h if you don't need fmt::format and fmt/format.h
+// otherwise.
+
 #include "format.h"
